Change data model, refactor logger/page-analysis
Implement a basic visit/page data model. Each new visit creates and refers
to a page doc, which can be deduplicated later if pages appear to be equal.

Let activity-logger only log the visits, and trigger a separate module,
page-analysis, to extract the page contents etc.

Put control in background instead of in content scripts. Use RPCs to perform
actions that must be executed in the content script.

Try harder to capture a screenshot by waiting until the tab is active and
completely loaded.