Reorganise searching&handling of pages and visits
Nest the referred page doc in each visit doc on retrieval.
(it might be nice to implement nesting linked docs in PouchDB itself..)

Pass the whole result object around (instead of putting only result.rows in
the overview state), for flexibility and consistency.

Rename WebPageAsListItem -> VisitAsListItem, because it represents a visit.