Create browser history import mechanics
Just create the mechanism, it is not yet invoked anywhere.

Puts each location/page/historyItem in the browser's history as a page in
our database. Gets each visitItem for each historyItem from the browser,
and puts it in our database as a visit, while mapping all references
between them to the corresponding new ids.

One remaining issue is deduplication. The new id of an imported item depends
on the timestamp+id of that item, so importing history is idempotent: items
that were already imported are ignored. However, if the a visit was also
logged by the activity logger, that visit will not be recognised as a
duplicate.