Store images as attachments & load them asyncly.
Rendering the image is the tricky part. We want to load it asynchronously,
which is the main point of using attachments. However, we cannot just give
the images their own URL and put that in the img src attribute, as we
cannot host files, and for a data/blob URI we have to have the attachment
loaded already.

(sidenote: in Firefox we could work around the hosting problem: we could
use made-up URLs, block requests to those URLs while returning a promise
that resolves with a {redirectUrl} containing the blob/data URI. But then
again, redirection to blob/data URIs is said not to work in e10s..)

The solution chosen here is to make a React component that fetches the
attachment, stores the data URI in its state, and passes it to its child
when it is available.
Using component state here, rather than Redux state, to hide this
irrelevant workaround from the rest of the application.

Possible improvements: generalise to not be specific to <img>, make it a
higher order component, perhaps using async-update-props or similar.