Replace usual search with a simple word filter.
pouchdb-quick-search was performing much too slow to be fun. Until we have
a faster alternative, we can use a quick rough hack: just filtering the
texts for presence of all query words. No indexing, stemming, tfidf, etc.

This takes time linear to the number of documents, but can easily be made
more efficient by limiting the number of visits we read, and loading more
of them when we need more results (= an infinite scroll solution).