Update overview UI to only req 1 result page at a time
Main change is how data is requested from search module (standard pagination append scheme):
- `limit` set to constant (page size)
- `skip` now derived from page size * page state

Sadly, because of the way `limit` is applied in `find-visits` in search module currently, this doesn't actually reduce amount of data requested from DB as the limit is applied after the query. But it sets up the UI to afford this if/when the search module changes in the future.

`currentPage` state also now defaults to 0 instead of 1 (initial skip should be 0 instead of page size).