Finish refactor, re-add loadMoreResults.
Adds more general cancellation possibilities. Just setting the 'exclude'
option did not permit e.g. a new search to cancel a running search caused
by scrolling down (= loadMoreResults).

It would be better to have waitingForResults be managed implicitly, and
read isPending info from the state, but that would require
asyncActionCreator to add logic to the reducer. Not sure how to do that
in a pretty way.

For now, the counter for waitingForResults is removed because there
cannot be multiple running searches in the current configuration.