Keep and compare current-&activeQueryParams.
Keep the last run query parameters, so we can check if we actually need
to run a new search when e.g. setQuery has occurred.

In particular this removes the annoying flicker when opening the
overview with a query (which runs the search directly, then triggers a
needless new search after the 500ms debounce time).