Properly merge contextual results.
The approach with uniqBy could cause the scenario that a hit in results
that are loaded later (when scrolling down) would be removed because
their id was already present; but this way they would still be displayed
small, as context results, while they should become real results.

There's probably smarter ways to go about. E.g., not adding context to
the end of the result list. But well, this works now.