Prevent deleting a 'higher fidelity' page.
Meaning, if a page had been freeze-dried, we only want to forget it in
favour of a new copy that also has a frozen version.

We could solve this in determinePageSameness, but it seems a slightly
different question: sameness is a symmetric test telling whether the
contents can be considered the same. We now also want to know, for two
same/similar documents, whether one has been better 'preserved' (like
two copies of the same book, of which one copy is more damaged).

We could choose to make the newer page redirect to the older one in case
the older one was of higher fidelity. Maybe for later.