Fixed bug which kept crashing the extension.
The CommentBox was programmed to autoresize as the user keeps typing, this was done by increasing
textarea's row if the textarea on the textarea's scroll event.

But when the textarea was resized manually it prevented adding the rows.

So this fix dynamically increases the height of the textarea instead of the rows.