Add total # annots to PageResultItem view
- super basic UI for now

Fix build errors

- handled some commits incorrectly

Fire a search with tag when a tag pill is clicked in Annotation Card.

Changed styling of “annotations on this page”

Added the number to be inside the comment bubble and added a new icon for the activated variant. Left out more styling work for when we merge ‘ribbon-sidebar’ into develop and then rebase this branch on develop, because the sidebar-ribbon branch is the authoriative styling work.

Fix content script crashing on load

- something with those methods that were being called every tab switch
- they attempt to toggle tooltip/ribbon on whatever tab, but apparently the new sidebar doesn't work like that
- also some weird import resolution issue in unrelated module which caused issues in the build output

Update content type filters UI

- no longer need annotations or website filter options, as we're splitting the searches
- only show the highlights + notes filters when annots search is on

Ensure search type switch shows in all search result views

- before wasn't being shown on invalid searches or 0 result

Fix typing + lint issues from split search changes

Hide annots appendix in page search results

Properly fix all annots search tests

- my attempt yesterday at this was really rushed, and I ended up not fixing things at all :/

Fix bug stopping sidebar from working

This was really painful. Turns out it was because a redux action
specific to the overview's redux store was imported here, although this
component is used in both options and content scripts.

Easiest work around was to do a conditional dynamic import if in options
script.

- also cleaned up annots list component - not directly related to the bug
but I ended up doing this in the process of finding the bug.

Fix sidebar crashing when opening collections menu

- another super hard to find bug
- turns out the `isList` prop was placed on the wrong component
- some other misc. code improvements that I was added while looking
for the bug