Fixes error thrown by AnnotList and TagList.
The array item passed to it was undefined.

Change the return structure of annots search.

Now, returns doc as the usual AnnotsPage array.
And sends an extra annotsByDay object.

Update rendering method and types in UI.

Annotation Search renders by using merging annotsByDay and SearchResults

Fix error which prevented fetching annotations.

Most of the frontend was still calling `getAllAnnotationsByUrl`.
But the remote function was renamed to `getAllAnnotations`.
Changed it back to `getAllAnnotationsByUrl` to solve this.