WIP Keyboard events on text inputs: remove redundent element path code
Originally was getting the HTML element in question from the event path
but any events fired on this component are for this text area so there's
no need to use it.

If for any reason events fire here that are not ment for the textarea
then can introduce it back in for disambiguation.