Backup Overlay refactor
This is an intermediary refactor, I've consolidated the old logic and made it typed, trying to make it more clear.
We should still move this the a UI Logic Class, hopefuly now it's a bit easier to make that move.

Included is also type local storage access
Included is also allowing PascalCase variable names, for use when defining react components.

Next up is the same treatment for the Backup Onboarding Wizard