Direct Linking - Remove it by default. Inc adding scaffolding to remove new features by install time but still be able to opt in to them.
- Needs fixing: The content tooltip width to react to 3 buttons not 4. The padding/margin of the inner components is not right.

Direct linking has been moved to be behind an opt in feature flag for now. It was decided in a conversation with @Oliver that it currently takes too long and is a feature that is not fully fleshed out yet, so in the interest of giving users simple features that work well, it's disabled for now.