Fix bookmark button in ribbon when bookmarking when ribbon is disabled.
There seems to be a race condition here between when (previously) oppertunisticaly toggling the bookmark state before saving it to the database, and the initialStore being loaded on first injection of the sidebar also setting the bookmark state.

To work around this I've set the bookmark state both before and after the action to save it. The state should always be correct now. If the save doesn't work the bookmark state is set back to what it was.

The trouble here hints at a much deeper cleanup of our store/state management needed at some point as part of ongoing refactoring efforts.