Search Results - delete modal enter key - allow Enter key to confirm deletion
This took a little longer than it should have. Due to react/modal usage the normal html autoFocus isn't reliable. Hence I've manually focused the button on component updates (modal opens). And changed the element from a div to a button so it captures the focus correctly. And reset the styles to account for this the browser HTML stylesheet.