Tag Picker: Add a bunch of tests and fix some bugs:
Fixed issues with adding/deleting tags in the replaced implementation on the popup and ribbon. (was using the confusing redux state actions which are seperate from the remote functions)
Fixed issues with which tags were returned in the results list depending on selected.
Added many UI Logic tests.
Refactored some logic to be simpler and easier to reason.