Tag Picker - Fix alltabs opacity issue. Mouse{Enter/Leave} error.
The onMouse{Enter/Over} handlers would work in the popup and storybook,
but not the ribbon.

The onMouse{Over/Out} handlers work in the ribbon too.

Why?

Does something in the ribbon already capture the onMouseEnter ? does
some other dom positioning or layout translate the position or steal the
focus?

This change seems to fix it.