Ensure Countly device_id is set to our internal user ID
I was confused about this. Previously it was set to a custom segmentation param, which still worked for grouping event counts between users but it means Countly would also assign its own ID to all exts (and possibly scripts). I know understand the device_id to be the unique user ID we've been talking about.