Implement pagination in sidebar searches
This took me almost a full day and the result is some really horrible untested code and a whole lot of stress :( It seems to work with the small set of sample data I was trying with, but no doubt will be littered with bugs.

I feel the majority of this pain is from the search endpoints we have, particularly with annotations search. Every different type of search can return different shaped results which need to be handled in different ways.
Searches also behave in different ways. eg: searching for annotations for a given page does not paginating on the `skip` param and instead expects `endDate` for various reasons, but only when there's no search terms. On search without terms it does.

There's probably very good reasons that things got this bad, which made sense at the times of implementation - particularly when just focused on that one small part of the code. Though I feel (when there's time to refactor search) we need to step back and look at all the different searches we have from a distance and unify them to behave and be interfaced with consistently.

NOTE: This is not the first time we (I) have had to deal with this pain with interacting with the searches. Though this is the first time I felt compelled enough to vent about it.
I'm also very aware that I was the one who designed and implemented most (all?) of the searches. Though I feel I did not really understand what I was doing, for example with the consequences of certain design choices, and there was never a chance for a second opinion or discussion about any of it.
I really feel like such an important and complicated part of the codebase like search needs attention and input from someone more experienced.