Annotation - Dashboard - Scroll to and open sidebar on click
When in the dashboard and showing the annotaions in it's sidebar.
Clicking on the go to annotation opens a new tab and in that tab scrolls
to the highlight in question and opens the sidebar showing the annotion
in it's list.

Required a hack due to content script in-tab functions loading after
page completion.