Performance - annotations - fix old attempt to wait for sidebar load
Porting the old functionality of clicking on an in-page highlight to
open that annotation in the sidebar included the call to attempt to wait
for the ribbon to load first. In the new refactored page ui
implementation, this is handled by the openSidebar or openRibbon itself
so it is redundant and harmful to do this within this function as it
was.