Reader - Refactoring in support of Reader annotation interactions.
Added a UILogic class that encapsulates the dashboard overview results.
Lifted the Sidebar and Viewer modal up to that so it can be controlled
through the `logic`. Currently broke the existing overview sideabr
implementation but will be easy to fix.

Fixed press X to close the reader modal.
Fixed sanitised parsing.