Refactoring to Annotation types for santiy and fixes
Most things back to working state, some work still on editing
Annotations and their effects showing immediately.

- Includes new faster storage read operation for getting annotations,
doing a fixed amount of queries and a linear amount of database work,
rather than an quadratic amount of queries and huge amount of databsae work.

- Re-Re-refactored the Annotation param uniqueAnnotationUrl back to just
url. It's non-obvious that it's actually the annotation's Primary Key,
but there's too much to massage and ends up inconsistant if changing it.
Really I'd like to see this changed in the schema, both naming and
representation.

- Misc fixes to do with annotation click events and content script
loading