Allow sidebar + ribbon to be reloaded more than once
This solves the issue with not being able to reload the ribbon+sidebar after closing them, though it's flawed as it introduces new problems.

Looking into the issues I noted:
- `src/content-scripts/content_scripts/global.ts` contains a `components` state keeping track of which in-page components are loaded
- this was added to in `SharedInPageUIState` when `loadComponent` was called, but it was never removed from when disabling the components
- so I added in a new `unloadComponent` function to pass into `SharedInPageUIState` to ensure that state is removed from when disabling those comps
- this means the next time they are attempted to be loaded, they actually load

This turned into a problem because AFAIK the special shadow DOM roots are never removed from the real DOM when
components get disabled. So every time you try to load a recently disabled component, it creates new dupe
elements in the real DOM to attach the shadow roots.
So I then:
- I found a method in `src/in-page-ui/dom`, `destroyRootElement` that actually handles removing those elements from the DOM which wasn't being used
- so I hooked that up to the code that handles "destroying" the sidebar and ribbon when being disabled
- this means the DOM now gets cleaned up when sidebar/ribbon is disabled