Revise state shape again so it's less nested
Really difficult for me to think about this. They last state shape was valid, but it proved to be a major pain to work with, trying to access pages and notes in nested arrays.
Spent some time reading some different ways to deal with similar shaped states, and this really helped with coming up with the new shape (read this years ago, but it was never so relevant until now):
https://redux.js.org/recipes/structuring-reducers/normalizing-state-shape