PDF Annotations - Indexing url & Anchoring
For prototyping purposes, have used a getUrl that checks if we're
viewing the pdf in our viewer (checking .endswith pdf)

ported in some more Hyp annotating code, fixed PDF anchors to be able to
work.

Still some timing issues - these logging statements actually seem to
slow the code down enough for it to work, without them it doesnt - Need
to put in beter checks that the PDFJS text layer hsa been rendered.

Also need to hook into when pages are created and destroyed from the
buffer to re-anchor.