PDF Annotations - Toggle Auto open PDFS
Since this is the old dashboard we're working with I've just copied the
logic layout of the Privacy setting for the settings toggle.

Since there's no way of telling the background script to completely
reload it's local storage settings, and we can't get the local storage
at page intercept time because it's async. I added an RPC method the
options page can call to reload the background PDF intercept settings. I
think this should be a global settings manager concern in the future.