RPC - Fixes static RPC facades ContentScript + New RPC
After the new RPC layer refactor, some things broke that should have not
been working in the first place.

In some cases the sidebar, ribbon and tooltip were creating their own instnace of the
RPC variable instead of using the one defined in the global content
script and passed down. This would cause them not to be initialised, it
also means that when this was working in the old arch, it was working
quite inefeciently and error prone.

I've refactored these calls to the remote functions to be passed down from
the global content script.